# docs/swagger.yaml
basePath: /
definitions:
...
  entities.BanListEntry:
    properties:
      banned_email:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
...
  errors.APIErrorResponse:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/errors.FieldError'
        type: array
      message:
        type: string
      meta: {}
      request_id:
        type: string
      status:
        type: integer
    type: object
  errors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
...
  requests.BanRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  responses.SimpleMessage:
    properties:
      data: {}
      message:
        example: Action completed successfully
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a comprehensive API for creating and managing appointments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asiko API
  version: "1.0"
paths:
...
  /ban-list:
    delete:
      consumes:
      - application/json
      description: Remove an email from the user's personal ban list.
      operationId: removeFromBanList
      parameters:
      - description: Email to unban
        in: body
        name: ban_request
        required: true
        schema:
          $ref: '#/definitions/requests.BanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SimpleMessage'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove email from ban list
      tags:
      - BanList
    get:
      description: Get a list of all emails on the user's personal ban list.
      operationId: getBanList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.BanListEntry'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's ban list
      tags:
      - BanList
    post:
      consumes:
      - application/json
      description: Add an email to the user's personal ban list.
      operationId: addToBanList
      parameters:
      - description: Email to ban
        in: body
        name: ban_request
        required: true
        schema:
          $ref: '#/definitions/requests.BanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.BanListEntry'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "409":
          description: Email already on ban list
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Add email to ban list
      tags:
      - BanList
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"